@@ -14,7 +14,7 @@ description: Aplicativo para conversa whatsapp sem necessidade de adição de co
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
@@ -22,9 +22,9 @@ dependencies:
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^0.1.2
-  country_pickers: ^1.2.0
-  url_launcher: ^5.4.2
+  cupertino_icons: ^1.0.3
+  country_pickers: ^2.0.0
+  url_launcher: ^6.0.10
 
 dev_dependencies:
   flutter_test:
